% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are.tolerably.equal.R
\name{are.tolerably.equal}
\alias{are.tolerably.equal}
\alias{are.tolerably.equal.default}
\title{Test For Equality With Tolerance}
\usage{
are.tolerably.equal(x, y, tol, ...)

\method{are.tolerably.equal}{default}(x, y, tol = .Machine$double.eps^0.5, ...)
}
\arguments{
\item{x}{A vector of numeric values}

\item{y}{Another vector of numeric values}

\item{tol}{The tolerance allowed between \code{x} and \code{y}
to consider them equal}

\item{...}{Ignored}
}
\value{
A logical vector with \code{TRUE} if the corresponding
        values in \code{x} and \code{y} are equal considering
        the specified tolerance level;
        \code{FALSE} otherwise
}
\description{
Determines if values are equals (within some tolerance).
}
\section{Methods (by class)}{
\itemize{
\item \code{are.tolerably.equal(default)}: Default implementation

}}
\examples{
are.tolerably.equal(2, 2.00001)
are.tolerably.equal(2, 2.00001, 0.001)
are.tolerably.equal(c(2, 2.01, 2.0001), 2, 0.001)

}
