% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.divisible.R
\name{is.divisible}
\alias{is.divisible}
\alias{is.divisible.default}
\title{Test For Divisibility With Tolerance}
\usage{
is.divisible(x, y, tol, ...)

\method{is.divisible}{default}(x, y, tol = .Machine$double.eps^0.5, ...)
}
\arguments{
\item{x}{A vector of numeric values with the dividends}

\item{y}{A vector of numeric values with the divisors}

\item{tol}{The tolerance allowed for \code{round(x / y)} from
whole numbers to consider \code{x} divisible by \code{y}}

\item{...}{Ignored}
}
\value{
A logical vector with \code{TRUE} if the corresponding
        value in \code{x} is divisible by the corresponding value
        in \code{y} considering the specified tolerance level;
        \code{FALSE} otherwise
}
\description{
Determines if some value(s) is(are) divisible by some other(s)
(within some tolerance).
}
\section{Methods (by class)}{
\itemize{
\item \code{is.divisible(default)}: Default implementation

}}
\examples{
is.divisible(10, 5)
is.divisible(10, 50)
is.divisible(10, 50, 0.5)
is.divisible(c(1, 5, 10), 50, 0.05)

}
