% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.wholenumber.R
\name{is.wholenumber}
\alias{is.wholenumber}
\alias{is.wholenumber.default}
\title{Test For Integrity With Tolerance}
\usage{
is.wholenumber(x, tol, ...)

\method{is.wholenumber}{default}(x, tol = .Machine$double.eps^0.5, ...)
}
\arguments{
\item{x}{A numeric value or a vector of numeric values}

\item{tol}{The tolerance allowed between \code{x} and
\code{round(x)} to consider \code{x} a whole number}

\item{...}{Ignored}
}
\value{
A logical vector with \code{TRUE} if the corresponding
        value in \code{x} is a whole number considering the
        specified tolerance level;
        \code{FALSE} otherwise
}
\description{
Determines if values are whole numbers (within some tolerance).
}
\section{Methods (by class)}{
\itemize{
\item \code{is.wholenumber(default)}: Default implementation

}}
\examples{
is.wholenumber(10 / 5)
is.wholenumber(10 / 50)
is.wholenumber(10 / 50, 0.5)
is.wholenumber(c(1, 5, 10) / 50, 0.05)

}
